Add DC Offset Control to Distortion
A parameter was added to implement a DC offset before the distortion
occurs. This control effectively adds even harmonics by increasing the
non-linearity of the distortion. The visible range is 0 to 100, but the
effective range is 0.0 to 0.5 as a positive additive to the signal.

In addition, a fixed 10Hz highpass filter was added after the distortion
stage to re-zero the signal's mean. This filter ensures the dc offset
only affects the distortion characteristic and doesn't output a nearly
0Hz tone.

Finally there is some simple refactoring to clean up repetitive left and
right usage and move them inside a loop for each channel.