Add Damping Knob to Reverb and Reduce Input Gain
I forgot to add the damping knob to the GUI when I first setup the
layout, so that was rectified. Also, I noticed that the wet signal was
significantly louder than the incoming dry signal, so I reduced the gain
by 0.5x before implementing the processor chain for the wet signal.