Refactor Everything using Best Practices
This commit affects every file in the repo. After watching some talks on
C++ best practices, as well as reading through the Core Guidelines, I
realized I had a lot to change. Everything that I had written worked,
but it was a collection of things I learned in C, or things I saw in
other people's code

Here is basically everything that I set out to do:

- Change all c-style casts to static_cast
- Change all c-style arrays to std::array
- Remove const function return types
- Set all referenced function arguments to const when possible
- Set all variables to const when possible
- Set object initialization type to auto when possible
- Remove unnecessary intermediate variables, especially filter coefficients
- Remove intermediate return values
- Remove dangling pointers
- Reduce the overuse of converting AudioBuffers to AudioBlock contexts
- Define constants, like number of channels or number of eq bands
- Clean up miscellaneous formatting