Refactor All Plugins for Readability
Reading Martin Fowler's book 'Refactoring' was the impetus for the last
few commits. The previous commits were specific to the Meter and GUI
components modules, and were essentially a test run. This commit sees
the same processes applied to all of the plugins.

The most frequent refactoring is extracting functions. The rule of thumb
that I used was if it needed a comment, it could be its own function.
This helped break long functions down into managable sections which
greatly aids readability.

The other changes are as follows:

- Using namespace juce in all module files
- Gain reduction vales were changed to arrays in every location
- All parameter values in modules are stored in structs instead of many
      global variables
- Filling buffers make use of 'makeCopyOf', which eliminates the need
      for channel iteration
- Replace conditional assignments with max and min where possible
- Remove superfluous comments